requirements.txt
numpy~=1.21.2
pillow~=8.4.0
onnxruntime~=1.10.0
torchvision~=0.11.3
streamlit~=1.5.1
gdown~=4.2.1